Move request.cb_kwargs tests to their own test file
