+++ tests/spiders.py
@@ -28,67 +28,6 @@ class MetaSpider(MockServerSpider):
         self.meta['close_reason'] = reason
 
 
-class KeywordArgumentsSpider(MockServerSpider):
-
-    name = 'kwargs'
-    checks = list()
-
-    def start_requests(self):
-        data = {'key': 'value', 'number': 123}
-        yield Request(self.mockserver.url('/first'), self.parse_first, cb_kwargs=data)
-        yield Request(self.mockserver.url('/general_with'), self.parse_general, cb_kwargs=data)
-        yield Request(self.mockserver.url('/general_without'), self.parse_general)
-        yield Request(self.mockserver.url('/no_kwargs'), self.parse_no_kwargs)
-        yield Request(self.mockserver.url('/default'), self.parse_default, cb_kwargs=data)
-        yield Request(self.mockserver.url('/takes_less'), self.parse_takes_less, cb_kwargs=data)
-        yield Request(self.mockserver.url('/takes_more'), self.parse_takes_more, cb_kwargs=data)
-
-    def parse_first(self, response, key, number):
-        self.checks.append(key == 'value')
-        self.checks.append(number == 123)
-        self.crawler.stats.inc_value('boolean_checks', 2)
-        yield response.follow(
-            self.mockserver.url('/two'),
-            self.parse_second,
-            cb_kwargs={'new_key': 'new_value'})
-
-    def parse_second(self, response, new_key):
-        self.checks.append(new_key == 'new_value')
-        self.crawler.stats.inc_value('boolean_checks')
-
-    def parse_general(self, response, **kwargs):
-        if response.url.endswith('/general_with'):
-            self.checks.append(kwargs['key'] == 'value')
-            self.checks.append(kwargs['number'] == 123)
-            self.crawler.stats.inc_value('boolean_checks', 2)
-        elif response.url.endswith('/general_without'):
-            self.checks.append(kwargs == {})
-            self.crawler.stats.inc_value('boolean_checks')
-
-    def parse_no_kwargs(self, response):
-        self.checks.append(response.url.endswith('/no_kwargs'))
-        self.crawler.stats.inc_value('boolean_checks')
-
-    def parse_default(self, response, key, number=None, default=99):
-        self.checks.append(response.url.endswith('/default'))
-        self.checks.append(key == 'value')
-        self.checks.append(number == 123)
-        self.checks.append(default == 99)
-        self.crawler.stats.inc_value('boolean_checks', 4)
-
-    def parse_takes_less(self, response, key):
-        """
-        Should raise
-        TypeError: parse_takes_less() got an unexpected keyword argument 'number'
-        """
-
-    def parse_takes_more(self, response, key, number, other):
-        """
-        Should raise
-        TypeError: parse_takes_more() missing 1 required positional argument: 'other'
-        """
-
-
 class FollowAllSpider(MetaSpider):
 
     name = 'follow'
